# OneDrive/Documents/Hackathon-master/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Event Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  location: string;
  category: string;
  capacity: number;
  registeredCount: number;
  attendedCount: number;
  imageUrl?: string;
  clubId?: number;
  clubName?: string;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  description: string;
  date: string;
  location: string;
  category: string;
  capacity: number;
  imageUrl?: string;
  clubId?: number;
}

export interface UpdateEventBody {
  title?: string;
  description?: string;
  date?: string;
  location?: string;
  category?: string;
  capacity?: number;
  imageUrl?: string;
}

export interface Registration {
  id: number;
  eventId: number;
  studentName: string;
  studentEmail: string;
  qrCode: string;
  attended: boolean;
  createdAt: string;
}

export interface CreateRegistrationBody {
  eventId: number;
  studentName: string;
  studentEmail: string;
}

export interface Attendance {
  id: number;
  registrationId: number;
  eventId: number;
  studentName: string;
  markedAt: string;
}

export interface MarkAttendanceBody {
  qrCode: string;
  eventId: number;
}

export interface Club {
  id: number;
  name: string;
  description: string;
  logoUrl?: string;
  category: string;
  memberCount: number;
  createdAt: string;
}

export interface CreateClubBody {
  name: string;
  description: string;
  logoUrl?: string;
  category: string;
}

export interface DashboardSummary {
  totalEvents: number;
  totalRegistrations: number;
  totalAttendance: number;
  totalClubs: number;
  upcomingEventsCount: number;
  attendanceRate: number;
}

export interface EventStat {
  eventId: number;
  eventTitle: string;
  registered: number;
  attended: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export type ListEventsParams = {
  category?: string;
  upcoming?: boolean;
  limit?: number;
};

export type GetEventFeaturedImage200 = {
  url: string;
};

export type ListRegistrationsParams = {
  eventId?: number;
  studentName?: string;
};

export type GetRegistrationQr200 = {
  qrCode: string;
  registrationId: number;
};

export type ListAttendanceParams = {
  eventId?: number;
};

export type GetUpcomingEventsParams = {
  limit?: number;
};