# OneDrive/Documents/Hackathon-master/lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Event Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendance {
  id: number;
  registrationId: number;
  eventId: number;
  studentName: string;
  markedAt: Date;
}