# OneDrive/Documents/Hackathon-master/lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Event Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: number;
  name: string;
  description: string;
  logoUrl?: string;
  category: string;
  memberCount: number;
  createdAt: Date;
}