# OneDrive/Documents/Hackathon-master/lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Event Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  eventId: number;
  studentName: string;
  studentEmail: string;
  qrCode: string;
  attended: boolean;
  createdAt: Date;
}