# OneDrive/Documents/Hackathon-master/lib/api-zod/src/generated/types/updateEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Event Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEventBody {
  title?: string;
  description?: string;
  date?: Date;
  location?: string;
  category?: string;
  capacity?: number;
  imageUrl?: string;
}